Moved `AriaRegion` to accessibility-utils
https://coveord.atlassian.net/browse/KIT-244
