+++ packages/atomic/src/components/atomic-no-results/atomic-no-results.tsx
@@ -11,7 +11,6 @@ import {
 } from '@coveo/headless';
 import {Component, h, Prop, State} from '@stencil/core';
 import {
-  AriaRegion,
   Bindings,
   BindStateToController,
   InitializeBindings,
@@ -19,6 +18,7 @@ import {
 import MagnifyingGlass from '../../images/magnifying-glass.svg';
 import escape from 'escape-html';
 import {Button} from '../common/button';
+import {AriaRegion} from '../../utils/accessibility-utils';
 
 /**
  * The `atomic-no-results` component displays search tips and a "Cancel last action" button when there are no results. Any additional content slotted inside of its element will be displayed as well.
