+++ packages/atomic/src/components/atomic-query-summary/atomic-query-summary.tsx
@@ -5,13 +5,13 @@ import {
   buildQuerySummary,
 } from '@coveo/headless';
 import {
-  AriaRegion,
   Bindings,
   BindStateToController,
   InitializableComponent,
   InitializeBindings,
 } from '../../utils/initialization-utils';
 import escape from 'escape-html';
+import {AriaRegion} from '../../utils/accessibility-utils';
 
 /**
  * The `atomic-query-summary` component displays information about the current range of results and the request duration (e.g., "Results 1-10 of 123 in 0.47 seconds").
