+++ packages/atomic/src/utils/accessibility-utils.tsx
@@ -1,0 +1,23 @@
+import {UpdateLiveRegionEventArgs} from '../components/atomic-search-interface/atomic-aria-live';
+import {buildCustomEvent} from './event-utils';
+import {InitializableComponent} from './initialization-utils';
+
+export function AriaRegion(regionName: string) {
+  function setMessage(message: string) {
+    document.dispatchEvent(
+      buildCustomEvent<UpdateLiveRegionEventArgs>(
+        'atomic/accessibility/updateLiveRegion',
+        {
+          region: regionName,
+          message,
+        }
+      )
+    );
+  }
+
+  return (component: InitializableComponent, setterName: string) => {
+    Object.defineProperty(component, setterName, {
+      set: (message) => setMessage(message),
+    });
+  };
+}
